Revamped and simplified DriveEngine
Relative is now based on the absolute system, but is still relative to Bogg's forward at the beginning of the motion. We no longer need a cumulative distance tracker, and we don't need to reset the encoders at any time, which means we don't need a discontinuity checker.

The checkpoint counter is also updated from a list of booleans to a double.

I removed the tangent power. I think some of the things I put in for safety end up working against us.